+++ libs/directus/directus-flexible-content/src/lib/components/attributes/index.ts
@@ -1,0 +1,9 @@
+// Add stack based remapping here
+
+const attributes: Record<string, string> = {
+  colspan: 'colSpan',
+  rowspan: 'rowSpan',
+  textAlign: 'textalign',
+}
+
+export default attributes