+++ libs/directus/directus-next-component/src/index.ts
@@ -1,3 +1,12 @@
 // Use this file to export React client components (e.g. those with 'use client' directive) or other non-server utilities
+export { default as DirectusLink } from './components/DirectusLink'
 
-export * from './lib/directus-next-component'
+export { default as useDirectusFile } from './hooks/directus-file'
+export { default as useDirectusLink } from './hooks/directus-link'
+export { default as useNavigationItems } from './hooks/navigation-items'
+
+export type { TFiles } from './types/files'
+export type { TPageSettings } from './types/page-settings'
+export type { TNavigationItems, TNavigationItemsTree } from './types/navigation-items'
+export type { TLinks } from './types/links'
+export type { TDirectusLinkProps } from './components/DirectusLink/interface'