+++ libs/stack/stack-ui/src/storybook/Lightbox/LightboxControlledState.tsx
@@ -1,0 +1,28 @@
+import { useState } from 'react'
+import Button from '../../components/Button'
+import type { TButtonProps } from '../../components/Button/interface'
+import Lightbox from '../../components/Lightbox'
+import type { TLightboxProps } from '../../components/Lightbox/interface'
+
+const LightboxControlledState = (props: TLightboxProps) => {
+  const { children, ...rest } = props
+
+  const [isOpen, setOpen] = useState(false)
+
+  const closeButtonRender = (buttonProps: TButtonProps) => (
+    <Button {...buttonProps} handlePress={() => setOpen(!isOpen)}>
+      Close lightbox
+    </Button>
+  )
+
+  return (
+    <>
+      <Button handlePress={() => setOpen(!isOpen)}>Open lightbox</Button>
+      <Lightbox {...rest} isOpen={isOpen} setOpen={setOpen} closeButton={closeButtonRender}>
+        {children}
+      </Lightbox>
+    </>
+  )
+}
+
+export default LightboxControlledState