+++ libs/stack/stack-ui/src/storybook/SelectField/index.tsx
@@ -1,5 +1,4 @@
 import { useState } from 'react'
-import { Item } from 'react-stately'
 import Select from '../../components/fields/Select/Select'
 import type { TSelectProps } from '../../components/fields/Select/Select.interface'
 
@@ -40,15 +39,17 @@ const SelectContent = ({
         {...rest}
         onSelectionChange={(key) => {
           setValue(key.toString())
+          // eslint-disable-next-line no-console
           console.log('selection changed')
         }}
         value={value}
-      >
-        <Item key={1}>Item One</Item>
-        <Item key={2}>Item Two</Item>
-        <Item key={3}>Item Three</Item>
-        <Item key={4}>Item Four</Item>
-      </Select>
+        options={[
+          { key: '1', value: 'Option 1' },
+          { key: '2', value: 'Option 2' },
+          { key: '3', value: 'Option 3' },
+          { key: '4', value: 'Option 4' },
+        ]}
+      />
     </>
   )
 }