+++ libs/stack/stack-ui/src/theme/index.tsx
@@ -3,6 +3,7 @@
 import React from 'react'
 import { makeTheme, createThemeProvider } from '../providers/Theme'
 import { accordionTheme } from './Accordion'
+import alertsTheme from './Alerts'
 import { container, grid, gridItem, main } from './Box'
 import button from './Button'
 import {
@@ -16,7 +17,7 @@ import {
   calendarNavigationButtonsContainer,
   calendarTable,
 } from './Calendar'
-import { carouselContainer, carouselNavigationButtons, carouselSlideContainer } from './Carousel'
+import carouselTheme from './Carousel'
 import { checkBox, checkMark, checkMarkIcon, checkboxContainer, checkboxLabel } from './Checkbox'
 import { checkboxGroup, checkboxGroupItemsGroup } from './CheckboxGroup'
 import {
@@ -30,9 +31,10 @@ import {
   datePickerLabel,
 } from './DatePicker'
 import { imgTheme } from './Image'
-import { lightBoxCloseBtn, lightBoxContainer, lightBoxLabel, lightBoxWrapper } from './LightBox'
+import lightboxTheme from './LightBox'
 import { radio, radioLabel, selectedMark } from './Radio'
 import { radioGroup, radioGroupItemsGroup } from './RadioGroup'
+import { searchIcon, searchInput } from './Search'
 import {
   shareButton,
   shareButtonContainer,
@@ -51,11 +53,7 @@ const BaseTheme = makeTheme({
   grid: (props) => grid(props),
   container: (props) => container(props),
   accordion: accordionTheme,
-  carousel: {
-    container: (props) => carouselContainer(props),
-    slideContainer: (props) => carouselSlideContainer(props),
-    navigationButtons: (props) => carouselNavigationButtons(props),
-  },
+  carousel: carouselTheme,
   calendar: {
     container: (props) => calendarContainer(props),
     header: (props) => calendarHeaderContainer(props),
@@ -88,12 +86,7 @@ const BaseTheme = makeTheme({
     container: (props) => sidePanelContainer(props),
     innerContainer: (props) => sidePanelInnerContainer(props),
   },
-  lightBox: {
-    wrapper: (props) => lightBoxWrapper(props),
-    container: (props) => lightBoxContainer(props),
-    closeBtn: (props) => lightBoxCloseBtn(props),
-    label: (props) => lightBoxLabel(props),
-  },
+  lightBox: lightboxTheme,
   shareButton: {
     icon: (props) => shareButtonIcon(props),
     button: (props) => shareButton(props),
@@ -146,13 +139,22 @@ const BaseTheme = makeTheme({
     input: (props) => textArea(props),
     errorMessage: (props) => typography({ ...props, size: 'footnotes', isError: true }),
   },
+  search: {
+    wrapper: () => 'min-w-fit sm:w-80 flex flex-col gap-2',
+    container: () =>
+      'relative flex items-center outline outline-2 outline-gray-300 focus-within:outline focus-within:outline-black focus-within:outline-2',
+    label: () => 'text-gray-3',
+    input: (props) => searchInput(props),
+    icon: (props) => searchIcon(props),
+    errorMessage: (props) => typography({ ...props, size: 'footnotes', isError: true }),
+  },
   textInput: {
     wrapper: () =>
-      `flex flex-col rounded-md px-4 py-1 mb-3 m-0.5 border-2 aria-disabled:pointer-events-none aria-disabled:opacity-30 focus-ring-black`,
-    label: () => 'text-xs',
+      `group flex flex-col rounded-md px-4 py-1 mb-3 m-0.5 border-2 aria-disabled:pointer-events-none aria-disabled:opacity-30 focus-ring-black`,
+    label: () => 'group-has-[:required]:after:content-["_*"] group-has-[:required]:after:text-red-500 text-xs',
     container: () => 'flex items-center gap-4',
     input: () => '',
-    errorMessage: (props) => typography({ ...props, size: 'footnotes', isError: true }),
+    errorMessage: (props) => typography({ ...props, className: 'ml-0', size: 'footnotes', isError: true }),
   },
   menu: {
     sidePanel: {
@@ -170,6 +172,7 @@ const BaseTheme = makeTheme({
     anchor: () => ``,
   },
   img: imgTheme,
+  alerts: alertsTheme,
 })
 
 export default React.memo(createThemeProvider(BaseTheme))